+++ owner-microservice/src/main/java/com/yamakuprina/kotiki/ownermicroservice/service/OwnersMQListener.java
@@ -21,7 +21,8 @@ public class OwnersMQListener {
 
     @RabbitListener(queues = MQConfig.OWNERS_QUEUE_ID)
     public OwnerDto findById(String id) {
-        return ownerService.findById(id);
+        OwnerDto ownerDto = ownerService.findById(id);
+        return ownerDto == null? new OwnerDto() : ownerDto;
     }
 
     @RabbitListener(queues = MQConfig.OWNERS_QUEUE_SAVE)
@@ -30,17 +31,18 @@ public class OwnersMQListener {
             ownerService.save(ownerDto);
             return "OK";
         } catch (Exception e) {
-            return "BAD_REQUEST";
+            return "500";
         }
     }
 
     @RabbitListener(queues = MQConfig.OWNERS_QUEUE_DELETE)
     public String delete(String id) {
         try {
+            if (ownerService.findById(id)==null) return "404";
             ownerService.delete(id);
             return "OK";
         } catch (Exception e) {
-            return "BAD_REQUEST";
+            return "500";
         }
     }
